requirements.txt
# Kimi Table to HTML 工具类依赖

# HTTP 请求库
requests>=2.25.0

# 环境变量管理
python-dotenv>=0.19.0

# HTML 解析库
beautifulsoup4>=4.9.0

# Selenium Web自动化
selenium>=4.0.0
